test: patch Cypress spec before any snapshots
