+++ __image_snapshots__/Authenticated Layout  should match image snapshot #0.png
