+++ cypress/support/commands.js
@@ -46,18 +46,3 @@ Cypress.Commands.add(
       });
   },
 );
-
-// @see https://github.com/meinaart/cypress-plugin-snapshots/issues/10
-export const fixCypressSpec = filename => () => {
-  const path = require('path');
-  const relative = filename.substr(1); // removes leading "/"
-  const projectRoot = Cypress.config('projectRoot');
-  const absolute = path.join(projectRoot, relative);
-  Cypress.spec = {
-    absolute,
-    name: path.basename(filename),
-    relative,
-  };
-};
-
-Cypress.Commands.add('fixSpec', fixCypressSpec);
