+++ cypress/support/index.js
@@ -19,3 +19,16 @@ import './commands';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
+
+// @see https://github.com/meinaart/cypress-plugin-snapshots/issues/10
+export const fixCypressSpec = filename => () => {
+  const path = require('path');
+  const relative = filename.substr(1); // removes leading "/"
+  const projectRoot = Cypress.config('projectRoot');
+  const absolute = path.join(projectRoot, relative);
+  Cypress.spec = {
+    absolute,
+    name: path.basename(filename),
+    relative,
+  };
+};
